+++ lizard_progress/parsers/tests/test_ribx_parser.py
@@ -1,4 +1,5 @@
 import datetime
+import mock
 
 import osgeo.ogr
 
@@ -11,7 +12,7 @@ from lizard_progress.tests.base import FixturesTestCase
 # Some helper variables for when the actual value doesn't matter.
 today = datetime.date.today()
 amersfoort = osgeo.ogr.Geometry(osgeo.ogr.wkbPoint)
-amersfoort.AddPoint(155000, 463000)
+amersfoort.AddPoint_2D(155000, 463000)
 
 
 class MockItem(object):
@@ -22,8 +23,12 @@ class MockItem(object):
         self.inspection_date = inspection_date
         self.geom = geom
         self.sourceline = sourceline
+        self.new = False
+        self.work_impossible = None
         if media is not None:
             self.media = media
+        else:
+            self.media = set()
 
 
 class TestSaveMeasurement(FixturesTestCase):
@@ -39,6 +44,16 @@ class TestSaveMeasurement(FixturesTestCase):
     def test_unknown_location_returns_none(self):
         item = MockItem('wrongref', today, amersfoort)
         self.assertEquals(self.parser.save_measurement(item), None)
+
+    def test_unknown_location_with_new_true_is_created(self):
+        # Parser needs a file_object, its name is used in the notification
+        file_object = mock.MagicMock()
+        file_object.name = 'testfile.ribx'
+        self.parser.file_object = file_object
+
+        item = MockItem('wrongref', today, amersfoort)
+        item.new = True
+        self.assertTrue(self.parser.save_measurement(item))
 
     def test_create_measurement_without_files(self):
         item = MockItem('testref', today, amersfoort, [])